refactor:
* change repo to jitpack
